feat: characters screen + alert dialog skeleton
