+++ frontend/lib/config/firebase_logic.dart
@@ -3,7 +3,11 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:role_maister/config/app_singleton.dart';
+import 'package:role_maister/models/aliens_character.dart';
+import 'package:role_maister/models/cthulhu_character.dart';
+import 'package:role_maister/models/dyd_character.dart';
 import 'package:role_maister/models/player.dart';
+import 'package:role_maister/widgets/widgets.dart';
 
 FirebaseService firebase = FirebaseService();
 
@@ -24,7 +28,7 @@ class FirebaseService {
   }
 
   // TODO: modificar esta función para el multiplayer
-  Future<Map<String, dynamic>> getCharacters(String gameId) async {
+  Future<Map<String, dynamic>> getCharactersFromGameId(String gameId) async {
     try {
       final DocumentReference gameReference =
           _firestore.collection("game").doc(gameId);
@@ -67,6 +71,90 @@ class FirebaseService {
     }
   }
 
+  Future<void> getAliensCharactersFromUserId(String userId) async {
+    List<AliensCharacter> characters = [];
+
+  try {
+    CollectionReference charactersCollection =
+        FirebaseFirestore.instance.collection('characters');
+
+    QuerySnapshot querySnapshot = await charactersCollection
+        .where('userId', isEqualTo: userId)
+        .get();
+
+    querySnapshot.docs.forEach((DocumentSnapshot document) {
+      Map<String, dynamic> data = document.data() as Map<String, dynamic>;
+
+      if (data.containsKey('aliensCharacters')) {
+        List<dynamic> charactersList = data['aliensCharacters'];
+
+        charactersList.forEach((characterData) {
+          AliensCharacter character = AliensCharacter.fromMap(characterData);
+          characters.add(character);
+        });
+      }
+    });
+  } catch (e) {
+    print("Error fetching characters: $e");
+  }
+}
+
+Future<void> getDydCharactersFromUserId(String userId) async {
+    List<DydCharacter> characters = [];
+
+  try {
+    CollectionReference charactersCollection =
+        FirebaseFirestore.instance.collection('characters');
+
+    QuerySnapshot querySnapshot = await charactersCollection
+        .where('userId', isEqualTo: userId)
+        .get();
+
+    querySnapshot.docs.forEach((DocumentSnapshot document) {
+      Map<String, dynamic> data = document.data() as Map<String, dynamic>;
+
+      if (data.containsKey('dydCharacters')) {
+        List<dynamic> charactersList = data['dydCharacters'];
+
+        charactersList.forEach((characterData) {
+          DydCharacter character = DydCharacter.fromMap(characterData);
+          characters.add(character);
+        });
+      }
+    });
+  } catch (e) {
+    print("Error fetching characters: $e");
+  }
+}
+
+Future<void> getCthulhuCharactersFromUserId(String userId) async {
+    List<CthulhuCharacter> characters = [];
+
+  try {
+    CollectionReference charactersCollection =
+        FirebaseFirestore.instance.collection('characters');
+
+    QuerySnapshot querySnapshot = await charactersCollection
+        .where('userId', isEqualTo: userId)
+        .get();
+
+    querySnapshot.docs.forEach((DocumentSnapshot document) {
+      Map<String, dynamic> data = document.data() as Map<String, dynamic>;
+
+      if (data.containsKey('cthulhuCharacters')) {
+        List<dynamic> charactersList = data['cthulhuCharacters'];
+
+        charactersList.forEach((characterData) {
+          CthulhuCharacter character = CthulhuCharacter.fromMap(characterData);
+          characters.add(character);
+        });
+      }
+    });
+  } catch (e) {
+    print("Error fetching characters: $e");
+  }
+}
+
   Future<bool> checkUsernameAlreadyExist(String username) async {
     try {
       final CollectionReference userReference =
@@ -94,13 +182,26 @@ class FirebaseService {
     }
   }
 
+Future<void> createRandomPlayer() async {
+    try {
+      AliensCharacter newRandomUser = AliensCharacter.random();
+      await firebase.createCharacter(newRandomUser.toMap());
+// TODO create random player
+      // Reload character data to update the UI
+    } catch (error) {
+      print("Error creating random player: $error");
+    }
+  }
+
   Future<void> saveUser(Player player) async {
     Map<String, dynamic> currentPlayer = {
       'uid': player.uid,
       'username': player.username,
       'email': player.email,
       'tokens': player.tokens,
-      'characters': [],
+      'aliensCharacters': player.aliensCharacters,
+      'dydCharacters': player.dydCharacters,
+      'cthulhuCharacters': player.cthulhuCharacters,
       'gamesPlayed': player.gamesPlayed,
       'experience': player.experience
     };
@@ -227,8 +328,7 @@ class FirebaseService {
       context.push("/rules");
       return user;
     } on FirebaseAuthException catch (e) {
-      if (e.code == 'email-already-in-use') {
-      }
+      if (e.code == 'email-already-in-use') {}
     } catch (e) {
       print(e);
     }
