+++ frontend/lib/models/models.dart
@@ -1,2 +1,2 @@
-export 'user_stats.dart';
+export 'aliens_character.dart';
 export 'chat_messages.dart';
