+++ frontend/lib/pages/register_page.dart
@@ -276,7 +276,7 @@ class _RegisterPageState extends State<RegisterPage> {
 
                     if (user != null) {
                       print("user not null");
-                      Player player = Player(uid: user.uid, username: username.text ,email: user.email, tokens: 5 ,characters: [], gamesPlayed: 0, experience: 1);
+                      Player player = Player(uid: user.uid, username: username.text ,email: user.email, tokens: 5 ,aliensCharacters: [], dydCharacters: [], cthulhuCharacters: [],gamesPlayed: 0, experience: 1);
                       // singleton.user = user;
                       singleton.player = player;
                       firebase.saveUser(player);
